Add insert() methods with support for bigger networks

+++ Network.h
@@ -17,12 +17,12 @@ public:
 	Network(Layer* new_layers);
 	void append(Layer new_layer);
 	void append(std::vector<Layer> new_layers);
-	void append(Layer* new_layers, int num_new_layers);
-	void insert(Layer new_layer, int position);
-	void insert(std::vector<Layer> new_layers, int position);
-	void insert(Layer* new_layers, int position);
+	void append(Layer* new_layers, long num_new_layers);
+	void insert(Layer new_layer, long position);
+	void insert(std::vector<Layer> new_layers, long position);
+	void insert(Layer* new_layers, long position);
 	void appendNetwork(Network new_network);
-	void insertNetwork(Network new_network, int position);
+	void insertNetwork(Network new_network, long position);
 };
 
 #endif
